Red Velvet Cake
Red velvet cake is traditionally a red, red-brown, crimson, or scarlet-colored chocolate layer cake, layered with ermine icing. Traditional recipes do not use food coloring, with the red color due to non-Dutched, anthocyanin-rich cocoa.

Common ingredients include buttermilk, butter, cocoa, vinegar, and flour. Beetroot or red food coloring may be used for the color.
History
Velvet cake is thought to have originated in Maryland in the early 20th century. In the 19th century, "velvet" cake, a soft and velvety crumb cake, came to be served as a fancy dessert, in contrast to what had been the more common, coarser-crumbed cake. Around the turn of the 20th century, devil's food cake was introduced, which is how some believe that red velvet cake came about. The key difference between the two cakes is that devil's food cake uses chocolate and red velvet cake uses cocoa.[5]

When foods were rationed in the US during World War II, bakers used boiled beetroot juices to enhance the color of their cakes. Beetroot is found in some red velvet cake recipes. Beetroot was and is used in some recipes as a filler or to retain moisture.[6] Adams Extract, a Texan company, is credited with bringing the red velvet cake to kitchens across America during the Great Depression era, by being one of the first to sell red food coloring and other flavor extracts with the use of point-of-sale posters and tear-off recipe cards.[7][8] The cake and its original recipe are well known in the United States from New York City's famous Waldorf-Astoria Hotel, which has dubbed the confection Waldorf-Astoria cake. However, it is widely considered a Southern recipe.[9] Traditionally, red velvet cake is iced with a French-style butter roux icing (also called ermine icing), which is very light and fluffy, but time-consuming to prepare. Cream cheese frosting and buttercream frosting are variations which have increased in popularity.[10][11]

In Canada, the cake was a well-known dessert in the restaurants and bakeries of the Eaton's department store chain in the 1940s and 1950s. Promoted as an exclusive Eaton's recipe, with employees who knew the recipe sworn to silence, many mistakenly believed the cake was the invention of the department store matriarch, Lady Eaton.[12]

In recent years, red velvet cake and red velvet cupcakes have become increasingly popular in the US and many European countries, especially around Christmas and more recently Valentine's Day. A resurgence in the popularity of this cake is attributed by some to the 1989 film Steel Magnolias, which included a red velvet groom's cake made in the shape of an armadillo.[9] Magnolia Bakery in Manhattan has served it since its opening in 1996, as did restaurants known for their Southern cooking like Amy Ruth's in Harlem,[9] which opened in 1998.[13] In 2000,[14] Cake Man Raven opened one of the first bakeries devoted to the cake in Brooklyn.[15]

Ingredients
Ingredients vary based on the era and area of the world. James Beard's reference, American Cookery (1972), describes three red velvet cakes varying in the amounts of shortening, butter, and vegetable oil. All used red food coloring. The reaction of acidic vinegar and buttermilk tends to better reveal the red anthocyanin in cocoa and keeps the cake moist, light, and fluffy. This natural tinting may have been the source for the name "red velvet", as well as "Devil's food" and similar names for chocolate cakes. Today, chocolate often undergoes Dutch processing, which prevents the color change of the anthocyanins. A reconstruction of the original red velvet cake involves reducing or eliminating the vinegar and colorants and using a non-Dutched cocoa to provide the needed acidity and color.

Variations
In addition to the many variations of red velvet cake, there are various red velvet-flavored products, including protein powder, tea, lattes, Pop-Tarts, waffles, and alcoholic beverages. The scent is used for candles and air fresheners as well.[16] For dietary restrictions, such as those due to allergies and ingredient sensitivity, vegan, gluten free, and dairy-free variations are available.[17]